Moved helpers out from under the integration directory.
